Removing some extra features that are not need for this demo:
* Removing the notifications that were generated when this application was
not running in the foreground.

* Remove the card service receiver that started the card service when the
device is powered on.

Add a new button to just dump the fingerprints container.
Change the button name for dumping the card containers.
Add a wait spinner dialog when the pkcs15 call is made as this might take
a while so let the user know something is happening.