Update application to support Sentry devices running Android Oreo (8) and Pie(9).
To accomodate changes required in Android for Oreo and Pie releases the
PCSC support library, libpcsclite.so, is platform depdendent. Therefore,
when building this application you must select the proper pcsc module
that corresponds to the Sentry Android version as follows:

Android OS	module
Lollipop	sentrypcsc-debug
Oreo 		sentrypcsc-2.0
Pie		sentrypcsc-3.0

You should modify you projects build and select one of the modules above
and rebuild the project for that platform.